Add Link For GoDiffX And View Project Opens In New Tab
GoDiffX project was given a link to my last demo video within
the playlist that has my previous 2. In order to do this, I revised
the code slightly so that instead of returning text, I returned a
JSX element with the proper text (and link) for all project descriptions.

Also changed 'View Project' button to open link in a new tab.